requirements.txt
openai
pandas
numpy
python-dotenv
gradio